remove improper use of InstallTool phase
Now skeletal dist files are created in the file gathering phase and edited in
the file munging stage; root files are created in AfterBuild.  The file
pruning operation is now detrimental as it would remove the file that we had
just added.

Use of proper phases now means that other plugins that care about the list of
files produced will find out about these files in time.  README is generated
via the file gathering phase, so it always exists by the time we need to fill
in the file's content.