Use main module pod file by default if available
If the main module has a companion pod file with the same basename, that
pod file will be used in preference to the main module file itself as
the source of the pod for the readme. This only affects the default and
has no effect if the source_filename option is specified explicitly in
the config.